A check has been added to ensure the notification shows when cart updates
An "IF" check has been added inside the useEffect hook this ensures the notification is only show when changes are made insde the cart and not when the app starts or is refreshed. To achieve this a new property called "changed" has been added to the "initialState" object inside cart-slice.